requirements.txt
alembic==0.9.8
aniso8601==3.0.0
attrs==19.1.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
coverage==4.4.2
dnspython==1.16.0
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-JWT-Extended==3.18.2
Flask-Limiter==1.0.1
Flask-Migrate==2.1.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gem==0.1.12
greenlet==0.4.15
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
limits==1.3
Mako==1.0.7
MarkupSafe==1.0
monotonic==1.5
passlib==1.7.1
pycparser==2.18
PyJWT==1.6.0
pyrsistent==0.15.1
python-dateutil==2.7.0
python-dotenv==0.10.2
python-editor==1.0.3
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2018.3
redis==3.2.1
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.25.2
Werkzeug==0.14.1
